requirements.txt
numpy==1.21.2
opencv-python-headless==4.5.5.64
torch==1.11.0

kafka-job-scheduler==1.2.*

# truncated requrements by YoloV5. See requirements.txt
matplotlib>=3.2.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
torchvision>=0.8.1
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012
psutil

seaborn>=0.11.0
tqdm>=4.41.0
pandas==1.3.4
